% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_helpers.R
\name{angle_helpers}
\alias{angle_helpers}
\alias{deg2rad}
\alias{rad2deg}
\alias{deg2vec}
\alias{vec2deg}
\alias{mean_vec}
\alias{mean_deg}
\title{angle helpers}
\usage{
deg2rad(x)

rad2deg(x)

deg2vec(x)

vec2deg(x)

mean_vec(x)

mean_deg(x)
}
\arguments{
\item{x}{test}
}
\description{
angle helpers
}
