% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_origin.R
\name{angle_point_origin}
\alias{angle_point_origin}
\title{angle_point_origin}
\usage{
angle_point_origin(interpol_grid_origin)
}
\arguments{
\item{interpol_grid_origin}{test}
}
\value{
test
}
\description{
angle_point_origin
}
