% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_helpers.R
\name{deg2rad}
\alias{deg2rad}
\alias{rad2deg}
\alias{deg2vec}
\alias{vec2deg}
\alias{rad2vec}
\alias{vec2rad}
\title{Angle transformation functions}
\usage{
deg2rad(x)

rad2deg(x)

deg2vec(x)

vec2deg(x)

rad2vec(x)

vec2rad(x)
}
\arguments{
\item{x}{Input for transformation
\itemize{
 \item{deg: }{Double. Angle in degrees}
 \item{rad: }{Double. Angle in radians}
 \item{vec: }{2 element double vector. Angle as unit vector}
}}
}
\value{
Same formats as x
}
\description{
Functions to transform angles from and to different formats.
The functions are not vectorised.
}
