% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_estimation_average_origin.R
\name{average_origin}
\alias{average_origin}
\alias{average_origin_searchid}
\alias{average_origin_searchid.default}
\alias{average_origin_searchid.mobest_origingrid}
\alias{average_origin_moving_window}
\alias{average_origin_moving_window.default}
\alias{average_origin_moving_window.mobest_origingrid}
\alias{no_data_windows}
\alias{no_data_windows.default}
\alias{no_data_windows.mobest_movingorigingrid}
\title{Summarize the results of the origin search}
\usage{
average_origin_searchid(origin_grid)

\method{average_origin_searchid}{default}(origin_grid)

\method{average_origin_searchid}{mobest_origingrid}(origin_grid)

average_origin_moving_window(
  origin_grid,
  window_start,
  window_stop,
  window_width,
  window_step
)

\method{average_origin_moving_window}{default}(
  origin_grid,
  window_start,
  window_stop,
  window_width,
  window_step
)

\method{average_origin_moving_window}{mobest_origingrid}(
  origin_grid,
  window_start,
  window_stop,
  window_width,
  window_step
)

no_data_windows(moving_origin_grid)

\method{no_data_windows}{default}(moving_origin_grid)

\method{no_data_windows}{mobest_movingorigingrid}(moving_origin_grid)
}
\arguments{
\item{origin_grid}{An object of class \code{mobest_origingrid} as created by
\link{search_spatial_origin}}

\item{moving_origin_grid}{An object of class \code{mobest_movingorigingrid}
as created by \link{average_origin_moving_window}}

\item{...}{undefined}
}
\value{
Different data products: \code{mobest_meanorigingrid},
\code{mobest_movingorigingrid} or \code{mobest_origingridnodatawindows}
}
\description{
Functions to transform the large origin grid objects to meaningful summary datasets
}
