% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpol_grid.R
\name{condense_interpol_grid}
\alias{condense_interpol_grid}
\title{condense_interpol_grid}
\usage{
condense_interpol_grid(interpol_grid)
}
\arguments{
\item{interpol_grid}{test}
}
\value{
test
}
\description{
condense_interpol_grid
}
