% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{create_kernel_grid}
\alias{create_kernel_grid}
\title{create_kernel_grid}
\usage{
create_kernel_grid(ds, dt, g)
}
\arguments{
\item{ds}{test}

\item{dt}{test}

\item{g}{test}
}
\description{
create_kernel_grid
}
