% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_basic_classes.R
\name{create_kernset_cross}
\alias{create_kernset_cross}
\title{create_kernel_grid}
\usage{
create_kernset_cross(ds, dt, g, on_residuals = T, auto = F)
}
\arguments{
\item{ds}{test}

\item{dt}{test}

\item{g}{test}
}
\description{
create_kernel_grid
}
