% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_basic_classes.R
\name{create_kernset_multi}
\alias{create_kernset_multi}
\title{Title}
\usage{
create_kernset_multi(d, g, on_residuals = T, auto = F, ..., it)
}
\arguments{
\item{it}{}
}
\value{

}
\description{
Title
}
