% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_grid.R
\name{create_model_grid}
\alias{create_model_grid}
\title{create_model_grid}
\usage{
create_model_grid(data, independent, dependent, kernel, prediction_grid)
}
\arguments{
\item{data}{test}

\item{independent}{test}

\item{dependent}{test}

\item{kernel}{test}

\item{prediction_grid}{test}
}
\description{
create_model_grid
}
