% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_grid.R
\name{create_model_grid_raw}
\alias{create_model_grid_raw}
\title{create_model_grid_raw}
\usage{
create_model_grid_raw(
  independent_tables,
  dependent_vars,
  kernel_settings,
  pred_grids
)
}
\arguments{
\item{independent_tables}{test}

\item{dependent_vars}{test}

\item{kernel_settings}{test}

\item{pred_grids}{test}
}
\value{
test
}
\description{
create_model_grid_raw
}
