% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_pred_grid.R
\name{create_prediction_grid}
\alias{create_prediction_grid}
\title{Create a spatial point grid}
\usage{
create_prediction_grid(area, spatial_cell_size)
}
\arguments{
\item{area}{An object of class \code{sf}. Polygons where the spatial grid should
be constructed}

\item{spatial_cell_size}{Numeric. Size of the output spatial grid cells in the unit of
\code{area}. See \code{?sf::st_make_grid} for more info}

\item{temporal_layers}{Numeric vector. Temporal layers of the requested spatiotemporal grid}
}
\value{
An object of class \code{mobest_spatialpositions}
}
\description{
Create a spatial point grid
}
