% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporal_observations.R
\name{create_spatpos_uncertain}
\alias{create_spatpos_uncertain}
\title{Title}
\usage{
create_spatpos_uncertain(id, x, y, z, it)
}
\arguments{
\item{name}{}
}
\value{

}
\description{
Title
}
