% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_crossvalidation.R
\name{crossvalidate}
\alias{crossvalidate}
\title{crossvalidate}
\usage{
crossvalidate(independent, dependent, kernel, iterations, groups = 10)
}
\arguments{
\item{independent}{test}

\item{dependent}{test}

\item{kernel}{test}

\item{iterations}{test}

\item{groups}{test}
}
\description{
crossvalidate
}
