% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{data_types}
\alias{data_types}
\alias{create_geopos}
\alias{create_prediction_grid}
\alias{create_geopos_multi}
\alias{geopos_to_spatpos}
\alias{create_spatpos}
\alias{create_spatpos_multi}
\alias{create_obs}
\alias{create_obserror}
\alias{create_obs_obserror}
\alias{create_obs_multi}
\alias{create_obs_obserror_multi}
\alias{create_kernel}
\alias{create_kernset}
\alias{create_kernset_multi}
\title{Input data type constructors}
\usage{
create_geopos(id, x, y, ...)

create_prediction_grid(area, spatial_cell_size)

create_geopos_multi(..., .names = NULL)

geopos_to_spatpos(geopos, z)

create_spatpos(id, x, y, z, ...)

create_spatpos_multi(..., .names = NULL)

create_obs(..., .names = NULL)

create_obserror(..., .names = NULL)

create_obs_obserror(obs, obserror)

create_obs_multi(..., .names = NULL)

create_obs_obserror_multi(..., .names = NULL)

create_kernel(dsx, dsy, dt, g, on_residuals = T, auto = F)

create_kernset(..., .names = NULL)

create_kernset_multi(..., .names = NULL)
}
\arguments{
\item{id}{Vector. Identifiers of the individual data points}

\item{x}{Numeric vector. Spatial x-axis coordinates}

\item{y}{Numeric vector. Spatial y-axis coordinates}

\item{...}{Different inputs - usually individual elements to be merged in a list type}

\item{area}{An object of class \code{sf}. Polygons where the spatial grid should
be constructed}

\item{spatial_cell_size}{Numeric. Size of the output spatial grid cells in the unit of
\code{area}. See \code{?sf::st_make_grid} for more info}

\item{.names}{Vector. Names of different object iterations}

\item{geopos}{An object of class `mobest_spatialpositions`.}

\item{z}{Numeric vector. Temporal positions}

\item{obs}{An object of class `mobest_observations`.}

\item{obserror}{An object of class `mobest_observations_error`.}

\item{dsx}{Double. Kernel lengthscale parameter for the x dimension (spatial x-axis). See \code{?laGP::newGP} for more information}

\item{dsy}{Double. Kernel lengthscale parameter for the y dimension (spatial y-axis)}

\item{dt}{Double. Kernel lengthscale parameter for the z dimension (temporal axis)}

\item{g}{Double. Kernel nugget parameter}

\item{on_residuals}{Logical. In the field calculation:
Should a linear model take over the main trends before the kriging interpolation?}

\item{auto}{Logical. In the field calculation:
Should the lengthscale and nugget values be automatically determined by laGPs
maximum likelihood algorithm? See \code{?laGP::mleGPsep} for more info}
}
\description{
Functions to create the main input data types for the mobest package.
}
