% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_estimation.R
\name{estimate_mobility}
\alias{estimate_mobility}
\alias{estimate_mobility.default}
\alias{estimate_mobility.mobest_origin_grid}
\title{estimate_mobility}
\usage{
estimate_mobility(origin_grid, mobility_regions)

\method{estimate_mobility}{default}(origin_grid, mobility_regions)

\method{estimate_mobility}{mobest_origin_grid}(origin_grid, mobility_regions)
}
\arguments{
\item{origin_grid}{Either an object of class \code{mobest_origin_grid} as
produced by \link{search_spatial_origin} or just an object of class
\code{mobest_interpol_grid}. Depending on the input a different mobility
estimation algorithm is applied}

\item{mobility_regions}{test}
}
\value{
test
}
\description{
estimate_mobility
}
