% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R, R/interpolation.R
\name{interpolate}
\alias{interpolate}
\alias{create_model_grid}
\alias{run_model_grid}
\title{3D interpolation with laGP}
\usage{
interpolate(
  independent,
  dependent,
  pred_grid,
  d = NA,
  g = NA,
  auto = F,
  on_residuals = T
)

create_model_grid(independent, dependent, kernel, prediction_grid)

run_model_grid(model_grid, unnest = T, quiet = F)

interpolate(
  independent,
  dependent,
  pred_grid,
  d = NA,
  g = NA,
  auto = F,
  on_residuals = T
)
}
\arguments{
\item{independent}{An object of class mobest_spatiotemporalpositions.
Spatiotemporal input point positions}

\item{dependent}{Numeric vector.
Dependent variable that should be interpolated}

\item{pred_grid}{An object of class mobest_spatiotemporalpositions.
Prediction grid positions for the interpolation}

\item{d}{Numeric vector. Lengthscale parameter.
See \code{?laGP::newGP} for more info}

\item{g}{Numeric. Nugget parameter}

\item{auto}{Should the lengthscale and nugget values be automatically determined
by laGPs maximum likelihood algorithm? See \code{?laGP::mleGPsep} for more info}

\item{on_residuals}{Should a linear model be wrapped around the kriging model
to handle the main trends independently?}

\item{kernel}{An object of class \code{mobest_kernelsetting_multi}
as created by \link{create_kernset_multi}. Kernel parameter settings.}

\item{prediction_grid}{An object of class \code{mobest_spatiotemporalpositions_multi}
as for example created by \link{create_prediction_grid}.
Prediction grid positions for the interpolation.}

\item{model_grid}{An object of class \code{mobest_modelgrid}
as created by \link{create_model_grid}.}

\item{unnest}{Logical. Should the kriging result be unnested to return a
prediction point-wise table of class \code{mobest_interpolgrid}?}

\item{quiet}{Logical. Should a progress indication be printed?}
}
\value{
Output of \code{?laGP::predGPsep}

Output of \code{?laGP::predGPsep}
}
\description{
Performs kriging with laGPs gaussian process prediction. See
\code{?laGP::predGPsep} for more information.

Construct and run a model grid with all combinations of the different input parameter
configurations for the kriging model.

Performs kriging with laGPs gaussian process prediction. See
\code{?laGP::predGPsep} for more information.
}
