% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_laGP.R
\name{interpolate_laGP}
\alias{interpolate_laGP}
\title{interpolate_laGP}
\usage{
interpolate_laGP(
  independent,
  dependent,
  pred_grid,
  auto = T,
  d,
  g,
  on_residuals = F
)
}
\arguments{
\item{independent}{test}

\item{dependent}{test}

\item{pred_grid}{test}

\item{auto}{test}

\item{d}{test}

\item{g}{test}

\item{on_residuals}{test}
}
\value{
test
}
\description{
interpolate_laGP
}
\examples{
\donttest{
independent <- tibble::tribble(
  ~x, ~y, ~z,
  1,1,2,
  3,2,1,
  2,3,3,
  7,7,7,
  8,7,9,
  9,7,8
)

dependent <- c(1,2,1,5,5,6)

pred_grid <- tibble::as_tibble(expand.grid(x = 1:10, y = 1:10, z = 1:10))

pred <- interpolate_laGP(independent, dependent, pred_grid, auto = F, d = c(3, 3, 4), g = 0.1, on_residuals = F)

pred_grid$pred_mean <- pred$mean

ggplot(data = pred_grid[pred_grid$z == 4, ]) +
  geom_raster(aes(x, y, fill = pred_mean))

}

}
