% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation}
\alias{interpolation}
\alias{create_model_grid}
\alias{run_model_grid}
\title{Create and run a kriging model grid}
\usage{
create_model_grid(independent, dependent, kernel, prediction_grid)

run_model_grid(model_grid, unnest = T, quiet = F)
}
\arguments{
\item{independent}{An object of class \code{mobest_spatiotemporalpositions_multi}
as created by \link{create_spatpos_multi}. Spatiotemporal input point positions.}

\item{dependent}{An object of class \code{mobest_observations_multi}
as created by \link{create_obs_multi}. Dependent variables that should be interpolated.}

\item{kernel}{An object of class \code{mobest_kernelsetting_multi}
as created by \link{create_kernset_multi}. Kernel parameter settings.}

\item{prediction_grid}{An object of class \code{mobest_spatiotemporalpositions_multi}
as for example created by \link{create_prediction_grid}.
Prediction grid positions for the interpolation.}

\item{model_grid}{An object of class \code{mobest_modelgrid}
as created by \link{create_model_grid}.}

\item{unnest}{Logical. Should the kriging result be unnested to return a
prediction point-wise table of class \code{mobest_interpolgrid}?}

\item{quiet}{Logical. Should a progress indication be printed?}
}
\description{
Constructs and run a model grid with all combinations of the different input parameter
configurations for the kriging model.
}
