% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_laGP_mle.R
\name{laGP_mle_anisotropic}
\alias{laGP_mle_anisotropic}
\title{Title}
\usage{
laGP_mle_anisotropic(
  independent,
  dependent,
  iterations,
  total_scaling_factor = 1000
)
}
\arguments{
\item{total_scaling_factor}{}
}
\value{

}
\description{
Title
}
