% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\alias{locate_multi}
\alias{multiply_dependent_probabilities}
\alias{sum_probabilities_per_group}
\title{Calculate a similarity/"origin" probability field for samples of interest}
\usage{
locate(
  independent,
  dependent,
  kernel,
  search_independent,
  search_dependent,
  search_space_grid,
  search_time = 0,
  search_time_mode = "relative",
  quiet = F
)

locate_multi(
  independent,
  dependent,
  kernel,
  search_independent,
  search_dependent,
  search_space_grid,
  search_time = 0,
  search_time_mode = "relative",
  quiet = F
)

multiply_dependent_probabilities(locate_overview, omit_dependent_details = T)

sum_probabilities_per_group(locate_product, ...)
}
\arguments{
\item{independent}{An object of class \code{mobest_spatiotemporalpositions_multi}
or \code{mobest_spatiotemporalpositions_multi} as created by \link{create_spatpos}
or \link{create_spatpos_multi}. Spatiotemporal input point positions to inform the
ancestry field.}

\item{dependent}{An object of class \code{mobest_observations} or
\code{mobest_observations_multi} as created by \link{create_obs} or
\link{create_obs_multi}. Dependent variables that should be interpolated for the
ancestry field.}

\item{kernel}{An object of class \code{mobest_kernelsetting} or
\code{mobest_kernelsetting_multi} as created by \link{create_kernset} or
\link{create_kernset_multi}. Kernel parameter settings for the ancestry fields.}

\item{search_independent}{An object of class \code{mobest_spatiotemporalpositions_multi}
or \code{mobest_spatiotemporalpositions_multi} as created by \link{create_spatpos}
or \link{create_spatpos_multi}. Spatiotemporal input point positions of the samples
of interest for which the similarity probability should be calculated. Must have
the same names as \code{independent}.}

\item{search_dependent}{An object of class \code{mobest_observations} or
\code{mobest_observations_multi} as created by \link{create_obs} or
\link{create_obs_multi}. Dependent variables of the samples of interest for which
the similarity probability should be calculated. Must have the same names as
\code{dependent}.}

\item{search_space_grid}{An object of class \code{mobest_spatialpositions}
as for example created by \link{create_prediction_grid}. Prediction grid positions
for which the similarity probabilities should be printed.}

\item{search_time}{Numeric vector. Time slices of the predition grid.}

\item{search_time_mode}{Character choice. One of "relative" or "absolute". Should
the search time be relative to the dating of the samples of interest or just absolute
points in time? The default for \code{search_time} and \code{search_time_mode} is
"relative" and 0, which causes the probabilities to be calculated for the exact dating
of the samples of interest.}

\item{quiet}{Logical. Should a progress indication be printed?}

\item{locate_overview}{An object of class \code{mobest_locateoverview} as created by
\link{locate} and \link{locate_multi}.}

\item{omit_dependent_details}{Logical. If TRUE, removes additional, dependent-wise
columns.}

\item{locate_product}{An object of class \code{mobest_locateproduct}.}

\item{...}{(Additional) grouping variables (\code{independent_table_id},
\code{dependent_setting_id}, \code{kernel_setting_id}, \code{pred_grid_id}, ...)}
}
\description{
Create a genetic ancestry field and determine similarity probabilities with it.
\link{locate_multi} handles permutations of input parameters, \link{locate} is
a simplified interface.
}
