% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_vectors.R
\name{determine_origin_vectors}
\alias{determine_origin_vectors}
\title{Search for a past point of "origin" in a prediction grid}
\usage{
determine_origin_vectors(locate_product, ..., quiet = F)
}
\arguments{
\item{locate_product}{An object of class \code{mobest_locateoverview_product} as created by
\link{locate} + \link{locate_multi} and \link{multiply_dependent_probabilities}.}

\item{...}{Additional grouping variables (\code{independent_table_id}, \code{dependent_setting_id},
\code{kernel_setting_id}, \code{pred_grid_id})}

\item{quiet}{Logical. Should a progress indication be printed?}
}
\value{
An object of class \code{mobest_origin_grid}
}
\description{
Search for a past point of "origin" in a prediction grid
}
