% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_distances.R
\name{calculate_pairwise_distances}
\alias{calculate_pairwise_distances}
\alias{calculate_geo_pairwise_distances}
\alias{calculate_time_pairwise_distances}
\alias{calculate_dependent_pairwise_distances}
\alias{bin_pairwise_distances}
\title{Calculate pairwise distances and prepare an empirical semivariogram}
\usage{
calculate_pairwise_distances(
  independent,
  dependent,
  m_to_km = T,
  with_resid = T
)

calculate_geo_pairwise_distances(independent, m_to_km = T)

calculate_time_pairwise_distances(independent)

calculate_dependent_pairwise_distances(
  ids,
  dependent,
  with_resid = F,
  independent = NULL
)

bin_pairwise_distances(
  pairwise_distances,
  geo_bin,
  time_bin,
  per_bin_operation = function(x) {     0.5 * mean(x^2, na.rm = T) }
)
}
\arguments{
\item{independent}{An object of class mobest_spatiotemporalpositions}

\item{dependent}{An object of class mobest_observations}

\item{m_to_km}{Logical. Should distances be transformed from m to km (x/1000)}

\item{with_resid}{Logical. Calculate distances also on the residuals of a spatiotemporal linear model}

\item{ids}{Vector. Sample identifiers (as in \code{mobest_spatiotemporalpositions}).}

\item{pairwise_distances}{An object of class mobest_pairwisedistances}

\item{geo_bin}{Numeric. Width of the spatial bins}

\item{time_bin}{Numeric. Width of the temporal bins}

\item{per_bin_operation}{Function. Summarising operation that should be applied to each distance.
Default: output = half mean squared input.}
}
\description{
Calculate pairwise distances and prepare an empirical semivariogram
}
