% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_laGP_mle.R,
%   R/parameter_estimation_variogram.R
\name{laGP_mle_sequence_isotropic_fixed_g}
\alias{laGP_mle_sequence_isotropic_fixed_g}
\alias{laGP_mle_anisotropic}
\alias{laGP_jmle_anisotropic}
\alias{calculate_pairwise_distances}
\alias{bin_pairwise_distances}
\title{Use laGPs mle algorithms for kriging kernel parameter estimation}
\usage{
laGP_mle_sequence_isotropic_fixed_g(
  independent,
  dependent,
  iterations,
  total_scaling_factor = 1000,
  g,
  space_time_scaling_factor_sequence,
  verb = 1
)

laGP_mle_anisotropic(
  independent,
  dependent,
  iterations,
  total_scaling_factor = 1000,
  verb = 1
)

laGP_jmle_anisotropic(
  independent,
  dependent,
  iterations,
  total_scaling_factor = 1000,
  verb = 1
)

calculate_pairwise_distances(independent, dependent, m_to_km = T)

bin_pairwise_distances(x, geo_bin = 100, time_bin = 100)
}
\arguments{
\item{independent}{An object of class mobest_spatiotemporalpositions}

\item{dependent}{An object of class mobest_observations}

\item{iterations}{Integer. Number of mle iterations}

\item{total_scaling_factor}{Numeric. It turned out that the parameter estimation
works better in certain numeric ranges. This scaling factor affects all three
dimensions}

\item{g}{Numeric. Fixed value for the nugget term}

\item{space_time_scaling_factor_sequence}{Numeric vector. Sequence of space-time
scaling factors}

\item{verb}{Integer. See \link[laGP]{mleGP} for more information}

\item{m_to_km}{Logical. Should distances be transformed from m to km (x/1000)}

\item{x}{An object of class mobest_pairwisedistances}

\item{geo_bin}{Numeric. Width of the spatial bins}

\item{time_bin}{Numeric. Width of the temporal bins}
}
\description{
Use laGPs mle algorithms for kriging kernel parameter estimation

Prepare an empirical semivariogram
}
