% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_pred_grid.R
\name{prediction_grid_for_spatiotemporal_area}
\alias{prediction_grid_for_spatiotemporal_area}
\title{Create a spatiotemporal point grid}
\usage{
prediction_grid_for_spatiotemporal_area(
  area,
  spatial_cell_size,
  temporal_layers
)
}
\arguments{
\item{area}{An object of class \code{sf}. Polygons where the spatial grid should
be constructed}

\item{spatial_cell_size}{Numeric. Size of the output spatial grid cells in the unit of
\code{area}. See \code{?sf::st_make_grid} for more info}

\item{temporal_layers}{Numeric vector. Temporal layers of the requested spatiotemporal grid}
}
\value{
Dataframe with columns x, y and z (spatiotemporal coordinates) and a
point identifier column id
}
\description{
Creates a prediction grid that can be used in the \link{create_model_grid} +
\link{run_model_grid} interpolation workflow
}
