% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_origin.R
\name{spatial_origin}
\alias{spatial_origin}
\title{spatial_origin}
\usage{
spatial_origin(interpol_grid, spatial_search_radius = 5e+05)
}
\arguments{
\item{interpol_grid}{test}
}
\value{
test
}
\description{
spatial_origin
}
