% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_variogram.R
\name{calculate_pairwise_distances}
\alias{calculate_pairwise_distances}
\alias{calculate_geo_pairwise_distances}
\alias{calculate_time_pairwise_distances}
\alias{calculate_dependent_pairwise_distances}
\alias{bin_pairwise_distances}
\title{Prepare an empirical semivariogram}
\usage{
calculate_pairwise_distances(
  independent,
  dependent,
  m_to_km = T,
  with_resid = T
)

calculate_geo_pairwise_distances(ids, independent, m_to_km = T)

calculate_time_pairwise_distances(ids, independent)

calculate_dependent_pairwise_distances(
  ids,
  dependent,
  with_resid = F,
  independent = NULL
)

bin_pairwise_distances(x, geo_bin = 100, time_bin = 100)
}
\arguments{
\item{independent}{An object of class mobest_spatiotemporalpositions}

\item{dependent}{An object of class mobest_observations}

\item{m_to_km}{Logical. Should distances be transformed from m to km (x/1000)}

\item{with_resid}{Logical. Calculate distances also on the residuals of a spatiotemporal linear model}

\item{ids}{Character vector. Identifier for the observations}

\item{x}{An object of class mobest_pairwisedistances}

\item{geo_bin}{Numeric. Width of the spatial bins}

\item{time_bin}{Numeric. Width of the temporal bins}
}
\description{
Prepare an empirical semivariogram
}
