% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_helpers.R
\name{mean_deg2vec}
\alias{mean_deg2vec}
\alias{cut_angle_deg}
\title{Angle helper functions}
\usage{
mean_deg2vec(x)

cut_angle_deg(x)
}
\arguments{
\item{x}{Double vector. Angles in degrees}
}
\value{
2 element double vector. Angle as unit vector
}
\description{
Functions for different operations on angles.
}
