% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_grid.R
\name{create_model_grid}
\alias{create_model_grid}
\title{create_model_grid}
\usage{
create_model_grid(independent_tables, dependent_vars, kernel_settings)
}
\arguments{
\item{independent_tables}{test}

\item{dependent_vars}{test}

\item{kernel_settings}{test}
}
\value{
test
}
\description{
create_model_grid
}
