% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_model_grid.R
\name{create_model_grid}
\alias{create_model_grid}
\title{Create a kriging model grid}
\usage{
create_model_grid(independent, dependent, kernel, prediction_grid)
}
\arguments{
\item{independent}{Named list of mobest_spatiotemporalpositions objects.
Spatiotemporal input point positions}

\item{dependent}{Named list of mobest_observations objects.
Dependent variables that should be interpolated. Each vector should have one
entry for each row in the \code{independent} list dataframes}

\item{kernel}{Named list of mobest_kernelsetting objects.
Kernel parameter settings. See \code{?interpolate_laGP} for more information#'}

\item{prediction_grid}{Named list of dataframes.
Prediction grid positions for the interpolation.
Each dataframe should have three numeric columns x, y and z and a point id column:

\itemize{
 \item{point_id: }{Unique point id}
 \item{x: }{Spatial coordinate in x-axis direction (in a cartesian grid)}
 \item{y: }{Spatial coordinate in y-axis direction}
 \item{z: }{Temporal position (age)}
}

See \code{?prediction_grid_for_spatiotemporal_area} for a function to create grid for a certain
spatial region}
}
\value{
An object of class \code{mobest_modelgrid} which inherits from tibble
}
\description{
Constructs a model grid with all combinations of the different input parameter
configurations for the kriging model.
}
