% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_model_grid.R
\name{create_model_grid}
\alias{create_model_grid}
\title{Create a kriging model grid}
\usage{
create_model_grid(independent, dependent, kernel, prediction_grid)
}
\arguments{
\item{independent}{Named list of dataframes. Spatiotemporal input point positions.
Each dataframe should have three numeric columns x, y and z:

\itemize{
 \item{x: }{Spatial coordinate in x-axis direction (in a cartesian grid)}
 \item{y: }{Spatial coordinate in y-axis direction}
 \item{z: }{Temporal position (age)}
}}

\item{dependent}{Named list of numeric vectors.
Dependent variables that should be interpolated. Each vector should have one
entry for each row in the \code{independent} list dataframes}

\item{kernel}{Named list of lists. Kernel parameter settings:

\itemize{
 \item{: }{Numeric vector with lengthscale values}
 \item{g: }{Nugget value}
 \item{on_residuals: }{Should a linear model take out the main trends before the kriging interpolation?}
 \item{auto: }{Should the lengthscale and nugget values be automatically determined by laGPs maximum likelihood algorithm?}
}

See \code{?interpolate_laGP} for more information}

\item{prediction_grid}{Named list of dataframes.
Prediction grid positions for the interpolation.
Each dataframe should have three numeric columns x, y and z and a point id column:

\itemize{
 \item{point_id: }{Unique point id}
 \item{x: }{Spatial coordinate in x-axis direction (in a cartesian grid)}
 \item{y: }{Spatial coordinate in y-axis direction}
 \item{z: }{Temporal position (age)}
 \item{region_id: }{Name of the region this point belongs to}
}

See \code{?create_prediction_grid} for a function to create grid for a certain
spatial region}
}
\value{
An object of class \code{mobest_modelgrid} which inherits from tibble
}
\description{
Constructs a model grid with all combinations of the different input parameter
configurations for the kriging model.
}
