% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_prep.R
\name{create_prediction_grid}
\alias{create_prediction_grid}
\title{create_prediction_grid}
\usage{
create_prediction_grid(
  area,
  spatial_cell_size = 1e+05,
  time_layers = seq(-7500, -500, 100)
)
}
\arguments{
\item{area}{test}

\item{spatial_cell_size}{test}

\item{time_layers}{test}
}
\value{
test
}
\description{
create_prediction_grid
}
