% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_grid.R
\name{create_prediction_grid}
\alias{create_prediction_grid}
\title{Create a spatiotemporal point grid}
\usage{
create_prediction_grid(
  area,
  mobility_regions,
  spatial_cell_size = 1e+05,
  time_layers = seq(-7500, -500, 100)
)
}
\arguments{
\item{area}{An object of class \code{sf}. Polygons where the spatial grid should
be constructed}

\item{mobility_regions}{An object of class \code{sf}. Polygons of regions the grid points
should be attributed to}

\item{spatial_cell_size}{Numeric. Size of the spatial grid cells in the unit of
\code{area}. See \code{?sf::st_make_grid} for more info}

\item{time_layers}{Numeric vector. Temporal layers of the requested spatiotemporal
grid}
}
\value{
Dataframe with columns x, y and z (spatiotemporal coordinates) and a
point identifier column point_id
}
\description{
Creates a prediction grid that can be used in the \link{create_model_grid} +
\link{run_model_grid} interpolation workflow#'
}
