% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_crossvalidation.R
\name{crossvalidate}
\alias{crossvalidate}
\title{Use crossvalidation for kriging kernel parameter estimation}
\usage{
crossvalidate(
  independent,
  dependent,
  kernel,
  iterations,
  groups = 10,
  quiet = F
)
}
\arguments{
\item{independent}{An object of class mobest_spatiotemporalpositions}

\item{dependent}{An object of class mobest_observations}

\item{kernel}{An object of class mobest_kernelsetting}

\item{iterations}{Integer. Number of crossvalidation iterations. Each iteration
goes along with a random reordering of the input points for training and test
data}

\item{groups}{Integer. Number of groups for splitting up training and test data
10 means for example that the data should be split up into 9 parts training and
1 part test observations}

\item{quiet}{Logical. Should a progress indication be printed?}
}
\description{
Use crossvalidation for kriging kernel parameter estimation
}
