% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_crossvalidation.R
\name{crossvalidate}
\alias{crossvalidate}
\title{crossvalidate}
\usage{
crossvalidate(
  independent,
  dependent,
  kernel,
  number_of_reorderings = 1,
  number_of_splits = 10
)
}
\arguments{
\item{independent}{test}

\item{dependent}{test}

\item{kernel}{test}

\item{number_of_reorderings}{test}

\item{number_of_splits}{test}
}
\description{
crossvalidate
}
