% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation.R
\name{crossvalidate}
\alias{crossvalidate}
\title{crossvalidate}
\usage{
crossvalidate(independent, dependent, kernel)
}
\arguments{
\item{independent}{test}

\item{dependent}{test}

\item{kernel}{test}
}
\description{
crossvalidate
}
