% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_estimation.R
\name{estimate_mobility}
\alias{estimate_mobility}
\title{estimate_mobility}
\usage{
estimate_mobility(spatial_origin, mobility_regions)
}
\arguments{
\item{spatial_origin}{test}

\item{mobility_regions}{test}
}
\value{
test
}
\description{
estimate_mobility
}
