% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mobility_estimation_mobility_measure_calculation.R
\name{estimate_mobility}
\alias{estimate_mobility}
\alias{estimate_mobility.default}
\alias{estimate_mobility.mobest_origin_grid}
\title{estimate_mobility}
\usage{
estimate_mobility(input_grid, delta_x, delta_y, delta_z)

\method{estimate_mobility}{default}(input_grid, delta_x, delta_y, delta_z)

\method{estimate_mobility}{mobest_origin_grid}(input_grid, delta_x, delta_y, delta_z)
}
\arguments{
\item{input_grid}{Either an object of class \code{mobest_input_grid} as
produced by \link{search_spatial_origin} or a specific object of class
\code{mobest_interpol_grid}. Depending on the input a different mobility
estimation algorithm is applied}

\item{delta_x}{test}

\item{delta_y}{test}

\item{delta_z}{test}
}
\value{
test
}
\description{
estimate_mobility
}
