% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_basic_classes.R
\name{create_obs}
\alias{create_obs}
\alias{create_spatpos}
\alias{create_spatpos_multi}
\alias{create_kernset}
\alias{create_kernset_multi}
\alias{create_kernset_cross}
\title{Input data type constructors}
\usage{
create_obs(...)

create_spatpos(id, x, y, z, ...)

create_spatpos_multi(id, x, y, z, it)

create_kernset(d, g, on_residuals = T, auto = F, ...)

create_kernset_multi(d, g, on_residuals = T, auto = F, ..., it)

create_kernset_cross(ds, dt, g, on_residuals = T, auto = F)
}
\arguments{
\item{id}{Vector. IDs of the observation points}

\item{x}{Numeric vector. Spatial x-axis coordinates}

\item{y}{Numeric vector. Spatial y-axis coordinates}

\item{z}{Numeric vector. Temporal positions}

\item{d}{3 element numeric vector. Kernel lengthscale parameter for the x, y and z dimension. See \code{?laGP::newGP} for more info}

\item{g}{numeric vector. Kernel nugget parameter}

\item{on_residuals}{logical. In the field calculation down the pipeline: Should a linear model take out the main trends before the kriging interpolation?}

\item{auto}{logical. In the field calculation down the pipeline:
Should the lengthscale and nugget values be automatically determined by laGPs
maximum likelihood algorithm? See \code{?laGP::mleGPsep} for more info}

\item{ds}{numeric vector. Different kernel lengthscale parameters for the x and y dimension
(isotropic)}

\item{dt}{numeric vector. Different kernel lengthscale parameters for the z dimension}
}
\value{
Different data types for specific applications.
}
\description{
Functions to create the main input data types for the mobest package.
See the README for example code.
}
