% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_basic_classes.R
\name{create_obs}
\alias{create_obs}
\alias{create_geopos}
\alias{create_geopos_multi}
\alias{create_spatpos}
\alias{create_spatpos_multi}
\alias{create_kernel}
\alias{create_kernset}
\alias{create_kernset_multi}
\title{Input data type constructors}
\usage{
create_obs(..., .names = NULL)

create_geopos(id, x, y, ...)

create_geopos_multi(..., .names = NULL)

create_spatpos(id, x, y, z, ...)

create_spatpos_multi(..., .names = NULL)

create_kernel(dsx, dsy, dt, g, on_residuals = T, auto = F)

create_kernset(..., .names = NULL)

create_kernset_multi(..., .names = NULL)
}
\arguments{
\item{...}{Different inputs (see examples in README)}

\item{.names}{Vector. Names of different object iterations}

\item{id}{Vector. IDs of the observation points}

\item{x}{Numeric vector. Spatial x-axis coordinates}

\item{y}{Numeric vector. Spatial y-axis coordinates}

\item{z}{Numeric vector. Temporal positions}

\item{dsx}{Double. Kernel lengthscale parameter for the x dimension (spatial x-axis). See \code{?laGP::newGP} for more info}

\item{dsy}{Double. Kernel lengthscale parameter for the y dimension (spatial y-axis)}

\item{dt}{Double. Kernel lengthscale parameter for the z dimension (temporal axis)}

\item{g}{Double. Kernel nugget parameter}

\item{on_residuals}{Logical. In the field calculation down the pipeline: Should a linear model take out the main trends before the kriging interpolation?}

\item{auto}{Logical. In the field calculation down the pipeline:
Should the lengthscale and nugget values be automatically determined by laGPs
maximum likelihood algorithm? See \code{?laGP::mleGPsep} for more info}
}
\value{
Different data types for specific applications.
}
\description{
Functions to create the main input data types for the mobest package.
See the README for example code.
}
