% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolate}
\alias{interpolate}
\title{3D interpolation with laGP}
\usage{
interpolate(
  independent,
  dependent,
  pred_grid,
  d = NA,
  g = NA,
  auto = F,
  on_residuals = T
)
}
\arguments{
\item{independent}{An object of class mobest_spatiotemporalpositions.
Spatiotemporal input point positions}

\item{dependent}{Numeric vector.
Dependent variable that should be interpolated}

\item{pred_grid}{An object of class mobest_spatiotemporalpositions.
Prediction grid positions for the interpolation}

\item{d}{Numeric vector. Lengthscale parameter.
See \code{?laGP::newGP} for more info}

\item{g}{Numeric. Nugget parameter}

\item{auto}{Should the lengthscale and nugget values be automatically determined
by laGPs maximum likelihood algorithm? See \code{?laGP::mleGPsep} for more info}

\item{on_residuals}{Should a linear model be wrapped around the kriging model
to handle the main trends independently?}
}
\value{
Output of \code{?laGP::predGPsep}
}
\description{
Performs kriging with laGPs gaussian process prediction. See
\code{?laGP::predGPsep} for more information.
}
