% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_laGP.R
\name{interpolate_laGP}
\alias{interpolate_laGP}
\title{interpolate_laGP}
\usage{
interpolate_laGP(independent, dependent, pred_grid, auto = T, d, g)
}
\arguments{
\item{independent}{test}

\item{dependent}{test}

\item{pred_grid}{test}

\item{auto}{test}

\item{d}{test}

\item{g}{test}
}
\value{
test
}
\description{
interpolate_laGP
}
