% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_laGP.R
\name{interpolate_laGP}
\alias{interpolate_laGP}
\title{interpolate_laGP}
\usage{
interpolate_laGP(
  independent,
  dependent,
  pred_grid,
  auto = F,
  d,
  g,
  on_residuals = T
)
}
\arguments{
\item{independent}{test}

\item{dependent}{test}

\item{pred_grid}{test}

\item{auto}{test}

\item{d}{test}

\item{g}{test}

\item{on_residuals}{test}
}
\value{
test
}
\description{
interpolate_laGP
}
\examples{
\donttest{
independent <- tibble::tribble(
  ~x, ~y, ~z,
  0,0,0,
  1,0,0,
  0,1,0,
  0,0,1,
  10,10,10,
  9,10,10,
  10,9,10,
  10,10,9
)

dependent <- c(1,1,1,1,10,10,10,10)

pred_grid <- tibble::as_tibble(expand.grid(x = 0:10, y = 0:10, z = 0:10))

pred <- interpolate_laGP(independent, dependent, pred_grid, auto = F, d = c(3, 3, 3)^2, g = 0.01, on_residuals = T)

pred_grid$pred_mean <- pred$mean

library(ggplot2)
ggplot(data = pred_grid) +
  geom_raster(aes(x, y, fill = pred_mean)) +
  facet_wrap(~z) +
  scale_fill_viridis_c()

}

}
