% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation_laGP_mle.R
\name{laGP_jmle_anisotropic}
\alias{laGP_jmle_anisotropic}
\title{Title}
\usage{
laGP_jmle_anisotropic(
  independent,
  dependent,
  iterations,
  scaling_factor = 1000
)
}
\arguments{
\item{scaling_factor}{}
}
\value{

}
\description{
Title
}
