% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_vectors.R
\name{origin_summary}
\alias{origin_summary}
\alias{pack_origin_vectors}
\alias{summarize_origin_vectors}
\alias{find_no_data_windows}
\title{Summarize the results of the origin search}
\usage{
pack_origin_vectors(origin_vectors, ...)

summarize_origin_vectors(
  origin_vectors,
  ...,
  window_start,
  window_stop,
  window_width,
  window_step
)

find_no_data_windows(origin_summary, ...)
}
\arguments{
\item{origin_vectors}{An object of class \code{mobest_originvectors} as created by
\link{determine_origin_vectors}}

\item{...}{(Additional) grouping variables (\code{independent_table_id}, \code{dependent_setting_id},
\code{kernel_setting_id}, \code{pred_grid_id}, ...)}

\item{window_start}{Start date of the moving window sequence}

\item{window_stop}{Stop date of the moving window sequence}

\item{window_width}{Width of each individual moving window}

\item{window_step}{Frequency of moving windows. Example:
If the first window starts at -3500 and extends until -3200, should the next
one start at -3450 or -3400, so with window_step = 50 or window_step = 100?}

\item{origin_summary}{An object of class \code{mobest_originsummary}
as created by \link{summarize_origin_vectors}}
}
\value{
\link{summarize_origin_vectors} returns and object of class
\code{mobest_originsummary}, \link{find_no_data_windows} then
\code{mobest_nodatawindows}.
}
\description{
Functions to transform the large origin vector tables to meaningful summaries.
\code{pack_origin_vectors} yields simple mean vectors,
\code{summarize_origin_vectors} a moving window summary through time.
}
