% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laGP_mle.R
\name{laGP_mle_sequence_isotropic_fixed_g}
\alias{laGP_mle_sequence_isotropic_fixed_g}
\alias{laGP_mle_anisotropic}
\alias{laGP_jmle_anisotropic}
\title{Use laGPs mle algorithms for kriging kernel parameter estimation}
\usage{
laGP_mle_sequence_isotropic_fixed_g(
  independent,
  dependent,
  iterations,
  total_scaling_factor = 1000,
  g,
  space_time_scaling_factor_sequence,
  verb = 1
)

laGP_mle_anisotropic(
  independent,
  dependent,
  iterations,
  total_scaling_factor = 1000,
  verb = 1
)

laGP_jmle_anisotropic(
  independent,
  dependent,
  iterations,
  total_scaling_factor = 1000,
  verb = 1
)
}
\arguments{
\item{independent}{An object of class mobest_spatiotemporalpositions}

\item{dependent}{An object of class mobest_observations}

\item{iterations}{Integer. Number of mle iterations}

\item{total_scaling_factor}{Numeric. It turned out that the parameter estimation
works better in certain (small) numeric ranges. This scaling factor affects all three
dimensions (x/scaling_factor)}

\item{g}{Numeric. Fixed value for the nugget term}

\item{space_time_scaling_factor_sequence}{Numeric vector. Sequence of space-time
scaling factors}

\item{verb}{Integer. See \link[laGP]{mleGP} for more information}
}
\description{
Use laGPs mle algorithms for kriging kernel parameter estimation
}
