% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_estimation_spatial_origin.R
\name{search_spatial_origin}
\alias{search_spatial_origin}
\title{Search for a past point of "origin" in a prediction grid}
\usage{
search_spatial_origin(
  interpol_grid,
  model_grid = NA,
  rearview_distance,
  cl = parallel::detectCores()
)

search_spatial_origin(
  interpol_grid,
  model_grid = NA,
  rearview_distance,
  cl = parallel::detectCores()
)

search_spatial_origin(
  interpol_grid,
  model_grid = NA,
  rearview_distance,
  cl = parallel::detectCores()
)
}
\arguments{
\item{interpol_grid}{An object of class \code{mobest_interpol_grid} as generated by
\link{run_model_grid}}

\item{rearview_distance}{How many years in the past should the reference point be
searched}

\item{cl}{A cluster object created by \link{parallel::makeCluster()},
or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel
evaluations (see \link{pbapply::pbapply()}).}
}
\value{
An object of class \code{mobest_origin_grid}
}
\description{
Search for a past point of "origin" in a prediction grid
}
