% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_estimation_spatial_origin.R
\name{search_spatial_origin}
\alias{search_spatial_origin}
\title{Search for a past point of "origin" in a prediction grid}
\usage{
search_spatial_origin(
  independent,
  dependent,
  interpol_grid,
  rearview_distance,
  quiet = FALSE
)
}
\arguments{
\item{independent}{An object of class mobest_spatiotemporalpositions.
Positions of the search points}

\item{dependent}{An object of class mobest_observations.
Observations at the search points}

\item{interpol_grid}{An object of class \code{mobest_interpolgrid} as generated by
\link{run_model_grid}}

\item{rearview_distance}{Numeric. How many years in the past should the reference point be
searched?}

\item{quiet}{Logical. Should a progress indication be printed?}
}
\value{
An object of class \code{mobest_origin_grid}
}
\description{
Search for a past point of "origin" in a prediction grid
}
