% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_origin_search.R,
%   R/mobility_estimation_spatial_origin.R
\name{locate}
\alias{locate}
\alias{multiply_dependent_probabilities}
\alias{search_spatial_origin}
\title{Search for a past point of "origin" in a prediction grid}
\usage{
locate(
  independent,
  dependent,
  kernel,
  search_independent,
  search_dependent,
  search_space_grid,
  search_time_distance = 0,
  quiet = F
)

multiply_dependent_probabilities(locate_overview, omit_dependent_details = T)

search_spatial_origin(
  independent,
  dependent,
  interpol_grid,
  rearview_distance,
  quiet = F
)
}
\arguments{
\item{independent}{An object of class mobest_spatiotemporalpositions.
Positions of the search points}

\item{dependent}{An object of class mobest_observations.
Observations at the search points}

\item{quiet}{Logical. Should a progress indication be printed?}

\item{interpol_grid}{An object of class \code{mobest_interpolgrid} as generated by
\link{run_model_grid}}

\item{rearview_distance}{Integer. How many years in the past should the reference point be
searched?}
}
\value{
An object of class \code{mobest_origin_grid}
}
\description{
Search for a past point of "origin" in a prediction grid
}
