% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_origin.R
\name{search_spatial_origin}
\alias{search_spatial_origin}
\alias{search_spatial_origin.default}
\alias{search_spatial_origin.mobest_interpol_grid}
\title{Search for a past point of "origin" in a prediction grid}
\usage{
search_spatial_origin(interpol_grid, steps = 3)

\method{search_spatial_origin}{default}(interpol_grid, steps = 3)

\method{search_spatial_origin}{mobest_interpol_grid}(interpol_grid, steps = 3)
}
\arguments{
\item{interpol_grid}{An object of class \code{mobest_interpol_grid} as generated by
\link{run_model_grid}}

\item{steps}{How many time steps in the past should the reference point be
searched}
}
\value{
An object of class \code{mobest_origin_grid}
}
\description{
Search for a past point of "origin" in a prediction grid
}
