% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_origin.R
\name{search_spatial_origin}
\alias{search_spatial_origin}
\title{search_spatial_origin}
\usage{
search_spatial_origin(interpol_grid)
}
\arguments{
\item{interpol_grid}{test}
}
\value{
test
}
\description{
search_spatial_origin
}
